Added migration.js to doc load
This is currently dummy file. My intention is to keep track of file format changes, and use this migration.js to detect, if running version of MaweJS can load the file, and what kinds of operations are needed to load it.

We might later have possibility to save files in specific formats.