Bug fix
It was not a good idea to disable the first item in the recent list. Yes, most of times it is the file already loaded to the editor. But if you close all files, you might want to load it.

So, we might need a bit more sophisticated method to determine, if you want to load the head of recent files, or not.